+++ school/src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -34,20 +34,7 @@ public class StudentController {
         this.studentService = studentService;
     }
 
-    @GetMapping("/avgAgeOfStudents")
-    public AvgAgeOfStudents getAvgAgeOfStudents() {
-        return studentService.getAvgAgeOfStudents();
-    }
 
-    @GetMapping("/countStudents")
-    public CountStudents getCountStudents() {
-        return studentService.getCountStudents();
-    }
-
-    @GetMapping("/fiveLastStudents")
-    public List<FiveLastStudents> getFiveLastStudents() {
-        return studentService.getFiveLastStudents();
-    }
     @PostMapping
     public Student writeStudent(@RequestBody Student student) {
         return studentService.writeStudent(student);
@@ -90,11 +77,6 @@ public class StudentController {
          return ResponseEntity.ok().build();
     }
 
-    @GetMapping("/avatars")
-    public ResponseEntity<List<Avatar>> getPagesOfAvatars(@RequestParam("page") Integer pageNumber, @RequestParam("size") Integer pageSize) {
-        List<Avatar> avatars = studentService.getAllAvatars(pageNumber, pageSize);
-        return ResponseEntity.ok(avatars);
-    }
     @PostMapping(value = "/{id}/avatar",consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<String> uploadAvatar(@PathVariable Long id, @RequestParam MultipartFile avatar) throws IOException {
         if (avatar.getSize() > 1024 * 300) {
