+++ school/src/main/java/ru/hogwarts/school/repository/StudentRepository.java
@@ -14,12 +14,6 @@ public interface StudentRepository extends JpaRepository<Student, Long> {
     List<Student> findByAge(int age);
 
     Collection<Student> findByAgeBetween(int min, int max);
-    @Query(value = "SELECT COUNT(*) FROM student", nativeQuery = true)
-    CountStudents countStudents();
 
-    @Query(value = "SELECT avg(age) as age FROM student ", nativeQuery = true)
-    AvgAgeOfStudents avgAgeOfStudents();
-    @Query(value = "SELECT * FROM student ORDER BY id DESC LIMIT 5", nativeQuery = true)
-    List<FiveLastStudents> fiveLastStudents();
 
 }
